+++ R/weightedVar.R
@@ -148,7 +148,9 @@ rowWeightedVars <- function(x, w = NULL, rows = NULL, cols = NULL,
   if (!is.null(w) && !is.null(cols)) w <- w[cols]
   
   # Preserve names attributes?
-  if (!(is.na(useNames) || useNames)) {
+  if (is.na(useNames)) {
+    deprecatedUseNamesNA()
+  } else if (!useNames) {
     rownames(x) <- NULL
   }
 
@@ -169,7 +171,9 @@ colWeightedVars <- function(x, w = NULL, rows = NULL, cols = NULL,
   if (!is.null(w) && !is.null(rows)) w <- w[rows]
   
   # Preserve names attributes?
-  if (!(is.na(useNames) || useNames)) {
+  if (is.na(useNames)) {
+    deprecatedUseNamesNA()
+  } else if (!useNames) {
     colnames(x) <- NULL
   }